Optimize pure-JS Draco decoder: ~1.3–1.5× faster, bit-exact with WASM, smaller bundle (#1)
Speeds up the hot paths of the pure-JS Draco decoder, makes its float math bit-exact with the C++/WASM reference, and trims the shipped bundle. Decoding is now **faster, byte-for-byte correct against the reference, and smaller to download.**

## Speed

Decode time vs. the previous `main`, same machine (`node tools/bench.mjs`):

| sample | before | this PR | speedup |
|---|--:|--:|:--:|
| duck.drc | 2.03 ms | 1.41 ms | 1.44× |
| bunny.drc | 9.80 ms | 6.34 ms | 1.55× |
| forest_house.glb | 5.25 ms | 4.20 ms | 1.25× |
| rolex.glb | 76.0 ms | 53.4 ms | 1.42× |
| ferrari.glb (51 prims) | 135.3 ms | 98.8 ms | 1.37× |

Real-world meshes decode **~1.25–1.55× faster** (tiny cubes ~1.1×).

## Correctness — bit-exact with WASM

Output is now **bit-identical to Google's `draco3d` WASM** across all **250** test primitives: `max float diff 0.00e+0 (eps 0)` (`node tools/verify-wasm.mjs`). Before this work, 247/250 diverged by up to ~6e-5 from float32 rounding.

Fixes behind the parity:
- Route quant/dequant and the octahedral transform through `Math.fround` to match C++ float32 arithmetic exactly (the core fix).
- Truncate the parallelogram WRAP prediction to int32 before clamping.
- Normalize `-0 → +0` in the octahedron normal rotations (keeps V8 on the Smi fast path; result-neutral).
- Guard against a null buffer in `PointAttribute.extractTo` (was crashing on empty attributes).

## Smaller bundle

Stronger minification (mangled private members + multi-pass terser) plus removal of decode-unreachable dead code:

| `build/DRACOLoader.min.js` | before | this PR |
|---|--:|--:|
| minified | 109.8 KB | 90.0 KB |
| gzipped | 24.8 KB | 23.7 KB |

The dead-code pass deletes ~1,300 lines the decoder never reaches — the `ValenceCache` subsystem, unused corner-table / transform / option / buffer methods, and unreachable branches — with byte-identical decode output, verified by the sha256 regression in `tools/bench.mjs`.

## What changed

- **Edgebreaker decoder** — inlined corner-table ops (`next` / `previous` / `opposite` / `swing*`) in the hottest loops; pre-allocated seam-corner arrays.
- **ANS entropy** — bit-shift instead of integer multiply in the renormalization loop.
- **Attribute extraction** — cached typed-array views + a contiguous-copy fast path.
- **Prediction schemes** — position caches and inlined helpers (parallelogram, geometric-normal, texcoord, octahedral).
- **Sequential mesh attribute decoding** — implemented, broadening format coverage.
- **Dead-code removal** — ~1,300 lines unreachable from the decode path deleted with byte-identical output.
- **Tooling** — `tools/bench.mjs` (decode timing + sha256 regression, optional `--wasm` JS× column) and `tools/verify-wasm.mjs` (parity vs `draco3d`).

Reproduce with `node tools/bench.mjs` and `node tools/verify-wasm.mjs`.

Live demo: https://raw.githack.com/Makio64/draco.js/31bc772b8f88439cacb43523aa6bfc6964d92d6b/index.html

+++ src/core/QuantizationUtils.js
@@ -9,17 +9,24 @@ export class Dequantizer {
 
   initFromRange(range, maxQuantizedValue) {
     if (maxQuantizedValue <= 0) return false;
-    this._delta = range / maxQuantizedValue;
+    // Match Draco C++ exactly (quantization_utils.cc): delta_ is a float and is
+    // computed as `range / static_cast<float>(max_quantized_value)` in float32.
+    // Doing the division in JS double and storing the result verbatim yields a
+    // value 1-2 ULP away from the WASM decoder for some attributes, so round
+    // every step to float32 with Math.fround.
+    this._delta = Math.fround(range / Math.fround(maxQuantizedValue));
     return true;
   }
 
   initFromDelta(delta) {
-    this._delta = delta;
+    this._delta = Math.fround(delta);
     return true;
   }
 
+  // C++ DequantizeFloat(int32 value): `value * delta_` evaluated in float32
+  // (the int is converted to float before the multiply).
   dequantizeFloat(val) {
-    return val * this._delta;
+    return Math.fround(Math.fround(val) * this._delta);
   }
 
   get delta() {
